+++ src/input/GpioButtonInput.cpp
@@ -7,7 +7,7 @@ GpioButtonInput::GpioButtonInput(GpioButtonMapping *button_mappings, size_t butt
     _button_count = button_count;
 
     // Initialize button pins.
-    for (int i = 0; i < _button_count; i++) {
+    for (size_t i = 0; i < _button_count; i++) {
         uint8_t pin = _button_mappings[i].pin;
         gpio::init_pin(pin, gpio::GpioMode::GPIO_INPUT_PULLUP);
     }
@@ -18,7 +18,7 @@ InputScanSpeed GpioButtonInput::ScanSpeed() {
 }
 
 void GpioButtonInput::UpdateInputs(InputState &inputs) {
-    for (int i = 0; i < _button_count; i++) {
+    for (size_t i = 0; i < _button_count; i++) {
         GpioButtonMapping button_mapping = _button_mappings[i];
         inputs.*(button_mapping.button) = !gpio::read_digital(button_mapping.pin);
     }
